Perform Differential Expression Analysis Using LIMMA

This function performs differential expression analysis on MS data using limma (Linear Models for Microarray Data) package. It supports a wide range of experimental designs, including handling of technical replicates, blocking factors, and the application of contrasts to identify differentially expressed proteins or peptides.

Usage

DE_stat_func(
  dat,
  sample = "SampleID",
  variable = "Protein",
  value = "Abundance",
  model.formula = "~ 0 + Condition",
  sva.obj = NULL,
  techrep = NULL,
  limma.trend = TRUE,
  limma.robust = TRUE,
  limma.winsor.tail.p = FALSE,
  time.course = FALSE,
  spline.df = 4,
  removeTotalMissing = TRUE,
  pairwise.contrasts = TRUE,
  pairwise.denominator = c("wt", "wildtype", "wild-type", "cntrl", "control", "ctl",
    "untreated"),
  manual.contrasts = NULL,
  choose.contrasts = FALSE,
  make.complex.contrasts = FALSE,
  blocking.parameter = NULL,
  reverse.contrasts = FALSE,
  showWarnings = FALSE,
  prefix = NULL,
  suffix = NULL,
  rmSingleShotProteins = TRUE,
  add.vs.inContrast = TRUE,
  path = getwd(),
  ...
)

Arguments

dat

A data.table containing the MS data. The table should include columns for samples, variables (e.g., proteins), and values (e.g., abundances).

sample

A character string specifying the column name that contains sample identifiers. Default is "SampleID".

variable

A character string specifying the column name that contains the variables of interest (e.g., "Protein"). Default is "Protein".

value

A character string specifying the column name that contains the values to be analyzed (e.g., "Abundance"). Default is "Abundance".

model.formula

A character string specifying the model formula to be used in the analysis. Default is "~ 0 + Condition".

sva.obj

An optional sva object for batch correction. Default is NULL.

techrep

A character string specifying the column name that contains technical replicates. If provided, these will be used as blocking factors. Default is NULL.

limma.trend

Logical or character. If TRUE, trend fitting is applied; if a character, it specifies the column name to be used for trend fitting. Default is TRUE.

limma.robust

Logical. If TRUE, robust fitting is applied in the eBayes step. Default is TRUE.

limma.winsor.tail.p

Logical or numeric. If FALSE, no winsorization is applied; if numeric, specifies the tail probability for winsorization. Default is FALSE.

time.course

Logical. If TRUE, a time course analysis is performed using spline regression. Default is FALSE.

spline.df

Integer. Degrees of freedom for spline regression in time course analysis. Default is 4.

removeTotalMissing

Logical. If TRUE, rows with all missing values are removed from the analysis. Default is TRUE.

pairwise.contrasts

Logical. If TRUE, all pairwise contrasts between conditions are computed. Default is TRUE.

pairwise.denominator

Character vector. Specifies the condition levels to be used as denominators in pairwise contrasts. Default includes "wt", "wildtype", "control", etc.

manual.contrasts

A list of manual contrasts to be applied. Default is NULL.

choose.contrasts

Logical. If TRUE, allows manual selection of contrasts from a list. Default is FALSE.

make.complex.contrasts

Logical. If TRUE, prompts the user to create complex contrasts. Default is FALSE.

blocking.parameter

A character string specifying the column name for blocking factors, typically technical replicates. Default is NULL.

reverse.contrasts

Logical. If TRUE, reverses the direction of the contrasts. Default is FALSE.

showWarnings

Logical. If TRUE, warnings are shown during execution. Default is FALSE.

prefix

A character string to prefix the output filenames. Default is NULL.

suffix

A character string to suffix the output filenames. Default is NULL.

rmSingleShotProteins

Logical. If TRUE, proteins with only one peptide are removed from the analysis. Default is TRUE.

add.vs.inContrast

Logical. If TRUE, the contrasts are labeled with "vs." between conditions. Default is TRUE.

path

A character string specifying the output directory for saving results. Default is the current working directory.

...

Additional arguments passed to the underlying LIMMA functions.

Value

A list containing:

"fit_eb": The fitted model after applying eBayes.

"fit_residuals": The residuals of the fit.

"fit0": The initial fitted model before applying contrasts.

Details

The DE_stat_func function is designed to handle complex experimental designs typical in mass spectrometry-based proteomics studies. It supports:

Technical Replicates: When techrep is provided, the function accounts for repeated measures by using the duplicateCorrelation method from LIMMA.

Time Course Analysis: If time.course is TRUE, the function performs spline regression to model the effect of time.

Contrast Definitions: Users can define contrasts manually or choose from pairwise contrasts. Complex contrasts, such as average or delta-delta contrasts, can also be generated.

This function is suitable for both differential expression analysis in standard proteomics experiments and more complex designs involving time courses or technical replicates. When working with experiments that have potential batch effects, users can incorporate surrogate variable analysis (sva.obj) to correct for these.