Generate result table of the Differential Expression Analysis

This function generates a differential expression statistics table from a fitted empirical Bayes model, merges it with additional data if provided, and optionally saves the results to a file.

Usage

DE_stat_table(
  fit_eb,
  dat,
  featureData = NULL,
  variable = "Protein",
  Gene.name = "Genes",
  adjust = "BH",
  save = TRUE,
  path = Tables.path,
  file.name = "StatTest_output",
  ...
)

Arguments

fit_eb

A fitted empirical Bayes model object (from limma).

dat

A data.table containing additional data to merge with the statistics table.

featureData

Optional data.table with feature data (gene names, ...) to be merged with the statistics table.

variable

A character string specifying the variable name for proteins or features in the data. Default is "Protein".

Gene.name

A character string specifying the column name for gene names. Default is "Genes".

adjust

A character string specifying the method for p-value adjustment. Default is "BH".

save

Logical, if TRUE, saves the result to a file. Default is TRUE.

path

A character string specifying the path to save the file. Default is Tables.path.

file.name

A character string specifying the file name for saving the output. Default is "StatTest_output".

...

Additional arguments passed to the export_DE_result function.

Value

A data.table containing the differential expression statistics.