Check Design of Experimental Data

This function verifies the design of experimental data provided in a data table. It checks for issues such as duplicate column names, multiple condition levels assigned to the same sample, and consistency in block and technical replicate assignments. Based on these checks, it determines the type of experimental design (e.g., fixed block, random block, or group comparison).

Usage

check_design(dat)

Arguments

dat

A data.table containing experimental design metadata. It should include columns for SampleID, Condition, and optionally Block and TechRep.

Value

A character string describing the type of experimental design detected and how the relevant factors are going to be handeled in the statistical model. Possible values include:

"fixBlock.fixTechRep" - block: fixed-effect, TechRep: fixed-effect

"fixBlock.randomTechRep" - block: fixed-effect, TechRep: random-effect

"randomBlock.avgTechRep" - block: random-effect, average over TechRep.

"fixBlock" - block: fixed-effect.

"randomBlock" - block: random-effect.

"fixTech" - TechRep: fixed-effect, no blocking.

"randomTech" - TechRep: random-effect, no blocking.

"group.comparison" - Simple group comparison design.

Examples

# Example data.table
dat <- data.table(SampleID = c("S1", "S2", "S3", "S4"),
                  Condition = c("C1", "C1", "C2", "C2"),
                  Block = c("B1", "B1", "B2", "B2"),
                  TechRep = c("T1", "T2", "T1", "T2"))
check_design(dat)
#> [1] "fixBlock.fixTechRep"