Check and Process Input File for MS Analysis

This function checks the input file or data table for MS analysis, determines the data level (PSM, peptide, or protein), identifies relevant columns, and processes the data to extract useful metadata.

Usage

check_input_file(x, sheet = 1)

Arguments

x

A data.table or a character string representing the path to an input file containing MS data. If a file path is provided, the file will be read into a data.table.

sheet

An integer specifying the sheet number to read from if the input file is an Excel file. Default is 1.

Value

A list with the following elements:

"MSmethod": The identified mass spectrometry method ("TMT" or "LFQ").

"data.level": The data level of the input data ("PSM-level", "Peptide-level", or "Protein-level").

"Intensity.col": The identified columns corresponding to intensity values.

"Sample.dt": A data.table containing metadata extracted from the input data, such as file IDs, channels, and conditions.

"parList": A list of identified relevant column names for further processing.

Details

The check_input_file function performs several key tasks:

Determines whether the data is at the PSM, peptide, or protein level based on the presence of specific columns.

Identifies the mass spectrometry method (TMT or LFQ) by examining the intensity columns.

Extracts relevant metadata, including sample IDs, conditions, and channels, and organizes it into a structured data.table.

Prepares a list of relevant column names for use in downstream analyses.

The function supports both direct data input as a data.table and reading data from an input file. It is designed to handle data from common MS analysis software (PD and MQ) outputs.

Examples

if (FALSE) { # \dontrun{
  # Check an input file for MS analysis
  input_file <- "path/to/ms_data.xlsx"
  result <- check_input_file(input_file, sheet = 1)
  print(result)
} # }