Clean Feature-level Data

This function cleans and filters feature-level (psm-level) data based on various criteria such as contaminant removal, isolation interference, and intensity thresholds. It also handles multiple PSM measurements per feature, removes single-shot proteins, and calculates basic statistics before and after filtering.

Usage

clean_feature_data(
  dat,
  metadata,
  filter.by.quaninfo = "auto",
  isolation.interference.cutoff = 75,
  collapse_psm_method = "TagPSM",
  filter.singleshot.proteins = "BYPEPTIDE",
  min.intensity = 0.01
)

Arguments

dat

A data.table containing the feature-level data.

metadata

A data.table containing metadata associated with the MS data.

filter.by.quaninfo

Character. Specifies the method for filtering MS data generated by PD based on quantitative information. Options are "auto", "none", or a specific label. Default is "auto".

isolation.interference.cutoff

Numeric or NULL. The threshold for filtering based on isolation interference. Features with interference above this value are removed. Default is 75.

collapse_psm_method

Character. Method for handling multiple PSM measurements per feature. Default is "TagPSM".

filter.singleshot.proteins

Character. Method for filtering single-shot proteins. Options are "BYPEPTIDE" or "BYFEATURE". Default is "BYPEPTIDE".

min.intensity

Numeric. Minimum intensity threshold below which intensity values are considered missing. Default is 0.01.

Value

A list containing the cleaned msDT, a BasicStats table with summary statistics, and Isol.Interference data.

Details

The function performs the following steps:

Removes rows where protein data is missing or marked as a contaminant.

Filters out single-shot proteins both before and after peptide filtering.

Removes features with intensities below the specified minimum threshold.

Filters features based on isolation interference and quantitative information.

Consolidates multiple PSM measurements per feature.

Outputs basic statistics before and after the cleaning process.

Examples

if (FALSE) { # \dontrun{
  clean_data <- clean_feature_data(dat = my_data, metadata = my_metadata)
} # }