Collapse Fractions in Mass Spectrometry Data

This function combines fractions in mass spectrometry data using various methods to summarize the intensities of features across fractions. The function supports multiple methods for summarizing the fractions, including maximum intensity, median intensity, sum of intensities, and tagging fractions.

Usage

collapse_fractions(
  dat = dat,
  group = "Pool",
  method = "MAXset",
  path = Tables.path
)

Arguments

dat

A data.table containing the MS data. The table should include columns such as Feature, Fraction, Condition if label-free, Channel if TMT, and Intensity.

group

A character string indicating the grouping variable for the analysis (e.g., "Pool"). Default is "Pool".

method

A character string specifying the method to use for combining fractions. Options include:

"MAXset": Select the fraction with the minimum missingness and maximum intensity across conditions/channels.

"MAX": Select the fraction with the maximum intensity across conditions/channels, regardless of missingness.

"MED" or "MEDIAN": Use the median intensity across fractions.

"SUM": Sum the intensities across fractions.

"SINGLE": Use only peptides with a single fraction and remove others.

"MULTI", "SEPERATE", "TAGFRAC", "TAGFRC": Tag features by fractions and pass to summarization.

Default is "MAXset".

path

A character string specifying the path to save the output summary tables. Default is Tables.path.

Value

A list with two elements:

"dat.fcomb": A data.table containing the combined fractions data.

"PrtPepCount_fraction": A data.table summarizing the protein and peptide counts before and after fraction combination.

Details

The function first identifies unique and non-unique measurements (features measured in more than one fraction per group). It then combines the fractions based on the specified method. The function can also generate a summary of the protein and peptide counts before and after fraction combination.