Generate a Configuration Template for MSqb2 Analysis

This function generates a configuration template for MSqb2 analysis, providing a set of predefined parameters that can be customized for specific mass spectrometry data processing tasks.

Usage

config_template()

Value

The function does not return any value but sets up a configuration template with predefined parameters. This template includes various settings for filtering, normalization, summarization, fraction combination, batch effect removal, missing data imputation, annotation, contrast definition, and reformatting of input data.

Details

The config_template function initializes a set of configuration parameters for MSqb2 analysis, which can be modified by the user as needed. The parameters are grouped into several categories:

Filters: Settings for filtering data based on quantification information, isolation interference, intensity thresholds, and more.

Normalization: Parameters for normalizing the data, including method selection and subset normalization.

Summarization: Options for summarizing data, including the method for median polishing.

Fraction Combination: Method for combining data across fractions.

Batch Effect Removal: Method for removing batch effects from the data.

Missing Data Imputation: Parameters for handling missing data.

Metadata Filtering: Options for filtering the dataset based on metadata.

Annotation: Source of feature annotation and the organism under study.

Contrasts: Parameters for defining contrasts in differential expression analysis, including pairwise contrasts and complex contrast definitions.

Reformatting: Options for reformatting input data, particularly for data generated from Proteome Discoverer.

The template is intended to be customized by users for their specific analysis needs, ensuring that all relevant parameters are set before running the MSqb2 pipeline.

Examples

if (FALSE) { # \dontrun{
  # Generate a configuration template
  config_template()

  # Users can modify the parameters as needed before running the analysis
} # }