Generate Quality Control Summary Tables

This function creates quality control (QC) summary tables from MS data and metadata. It calculates basic statistics, missingness per fraction and channel, and checks for feature overlap across fractions. The results can be saved to an Excel file and returned as a list for further analysis.

Usage

create_qc_tables(
  dat = dat,
  metadata = metadata,
  ms.software = "PD",
  save = TRUE,
  Tables.path = Tables.path
)

Arguments

dat

A data.table containing the mass spectrometry data. This table should include columns such as Feature, Protein, Peptide, Pool, Channel, Fraction, and Intensity.

metadata

A data.table containing metadata associated with the mass spectrometry data. This table may include columns such as Pool, Channel, and Fraction.

ms.software

A character string specifying the mass spectrometry software used, either of "PD" and "MQ". Default is "PD" (Proteome Discoverer).

save

Logical. If TRUE, the QC tables are saved to an Excel file. Default is TRUE.

Tables.path

A character string specifying the directory where the output file will be saved. Default is the value of Tables.path.

Value

A list containing the QC summary tables generated during the analysis, which may include:

"PSMsperFraction": A table showing the overlap of features across fractions.

"MissingPerFracChnl": A table summarizing the missingness of data per fraction and channel.

"extMixFrac": External mix fractions data (if applicable).

Details

The create_qc_tables function generates several summary tables that are essential for quality control in mass spectrometry-based proteomics studies:

Fractions Overlap: Evaluates how features are distributed across multiple fractions, particularly useful for TMT or LFQ methods.

Missingness Analysis: Provides a summary of missing data points per fraction and channel.

The function merges the MS data with the provided metadata, computes the QC statistics, and optionally saves the results in an Excel file for easy review and further analysis.