Check and Validate File Paths

This function verifies the existence of files specified in the input list. It checks if these files are either full paths that exist, or if they can be found in a specified directory (dpth) or the working directory. If files are missing, the function logs an error or warning and optionally copies the files to the dpth directory.

Usage

.check.file(input, dpth = Data.path, toData = FALSE, interactive = TRUE)

Arguments

input

A named list where names are file names or patterns and values are file paths. If a value is NULL, it will be ignored.

dpth

A character string specifying the directory path to search for files. Default is Data.path.

toData

A logical value indicating whether to copy files to the dpth directory if they are not already there. Default is FALSE.

interactive

A logical value indicating whether the function should prompt the user interactively to copy files. Default is TRUE.

Value

A named list of files with their paths. If files were copied, their paths in the dpth directory are returned.

Details

The function first checks if the files in input are full paths and exist. If so, it updates the output list and sets toData to TRUE. If not, it searches for these files in the specified dpth directory and, if necessary, in the working directory. If files are still not found, it logs a fatal message. If toData is TRUE and interactive is TRUE, the user will be prompted to copy missing files to the dpth directory.

Examples

if (FALSE) { # \dontrun{
# Example usage
input_files <- list(file1 = "path/to/file1.csv", file2 = "path/to/file2.csv")
output_files <- .check.file(input = input_files, dpth = tempdir(), toData = TRUE, interactive = TRUE)
print(output_files)
} # }