Import Configuration Parameters from a File

Imports configuration parameters from a specified file or generates a new configuration file if none exists. This function is typically called from msqb_config and supports both workflow and visualization configurations.

Usage

.importConfigfile(
  conf.fl,
  Scripts.path,
  analysis.name,
  add.date.tag,
  whichConf,
  conf.args
)

Arguments

conf.fl

A character string specifying the path to the configuration file. If NULL, a new configuration file is generated.

Scripts.path

A character string specifying the path where the scripts and configuration files are located.

analysis.name

A character string specifying the analysis name to include in the configuration file name. Default is NULL.

add.date.tag

A character string specifying a date tag to append to the configuration file name. Default is NULL.

whichConf

A character string specifying the type of configuration ("workflow" or "viz").

conf.args

A list of additional configuration arguments that can override parameters in the configuration file.

Value

A character string specifying the path to the imported or generated configuration file.

Details

If the specified configuration file does not exist or is NULL, the function creates a new configuration file with default parameters in the specified Scripts.path. The function reads the configuration file and evaluates its content in the parent environment.