Export Objects to Parent Environment

This function exports all objects from the current environment to the parent environment. It retrieves the objects from the current environment and places them into the environment that is one level up from the current environment.

Usage

expTopEnv()

Details

The function uses mget to get the objects from the current environment and list2env to export them to the parent environment. This is useful when you need to make objects available to a higher-level scope.

Examples

# Example usage:
{
  a <- 1
  b <- 2
  # a and b are in the current environment
  expTopEnv()
  # a and b should now be available in the parent environment
  print(parent.frame()$a) # Should print 1
  print(parent.frame()$b) # Should print 2
}
#> [1] 1
#> [1] 2