Impute Missing Values in Data

This function imputes missing values in a data set using various imputation methods. It supports a range of imputation techniques including K-nearest neighbors (KNN), maximum likelihood estimation (MLE), quantile regression imputation for left-censored data (QRILC), and hybrid methods combining different strategies for missing at random (MAR) and missing not at random (MNAR) data.

Usage

impute_missing_values(
  dat,
  val = "Abundance",
  col.p = "Filename",
  row.p = "Feature",
  na.imputation.method = "v2-mnar",
  hybrid.mar = "KNN",
  hybrid.mnar = "QRILC",
  n_knn = 15,
  tune.sigma = 0.5,
  q = 0.01,
  ...
)

Arguments

dat

A data.table containing the data to be imputed. The table should include columns corresponding to features, samples, and values to be imputed.

val

A character string specifying the column name that contains the values to be imputed. Default is "Abundance".

col.p

A character string specifying the column name(s) that either contains unique sample names (e.g., "Filename"), or a combination of which can be used to uniqule address each sample (e.g., c("Pool", "Channel") for MS method TMT). Default is "Filename".

row.p

A character string specifying the column name that contains the feature identifiers. Default is "Feature".

na.imputation.method

A character string specifying the method to use for imputation. Options include:

"V2": method v2 of the msImpute package

"V2-MNAR": method v2-mnar of the msImpute method specifically for MNAR data.

"QRILC": Quantile Regression Imputation of Left-Censored data.

"KNN": K-nearest neighbors.

"MLE": Maximum Likelihood Estimation.

"MINPROB": Impute using the minimal probability.

"MINDET": Impute using the minimal detection limit.

"ZERO": Impute with zeros.

"HYBRID": Hybrid method combining MAR and MNAR strategies.

Default is "v2-mnar".

hybrid.mar

A character string specifying the method to use for MAR imputation when using the hybrid approach. Default is "KNN".

hybrid.mnar

A character string specifying the method to use for MNAR imputation when using the hybrid approach. Default is "QRILC".

n_knn

An integer specifying the number of nearest neighbors to use for KNN imputation. Default is 15.

tune.sigma

A numeric value specifying the tuning parameter for the sigma in QRILC and MinProb methods. Default is 0.5.

q

A numeric value specifying the quantile for MinDet and MinProb methods. Default is 0.01.

...

Additional arguments passed to the imputation functions.

Value

A data.table containing the original data with missing values imputed.

Details

The impute_missing_values function offers several methods for imputing missing data, depending on the nature of the missingness and the experimental design. For in-depth details please refere to vignette of the packages msImpute and imputeLCMD.

"v2": This method, provided by the msImpute package, is designed for missing at random (MAR) data. It is particularly suitable for DIA experiments where missing values are assumed to be random.

"v2-mnar": This method extends "v2" to handle missing not at random (MNAR) data, making it suitable for cases where missing values are more likely for low-abundance features. This is often the case in DDA (like TMT) experiments, where the absence of data might be systematically related to the intensity of the signal.