Median Normalization of Data

This function performs median normalization on a dataset. Optionally, it can log2 transform the data before normalization and perform normalization based on a subset of the data specified by the user.

Usage

median_normalisation(
  dat,
  col.p,
  row.p,
  val,
  log2transform = TRUE,
  normalize.bySubset = NULL,
  normalize.bySubset.matchTerm = NULL,
  normalize.bySubset.matchTerm.exactMatch = FALSE
)

Arguments

dat

A data.table containing the data to be normalized. The table should include columns for the features, samples, and values to be normalized.

col.p

A character string specifying the column name that contains the sample identifiers (e.g., "SampleID").

row.p

A character string specifying the column name that contains the feature identifiers (e.g., "Protein").

val

A character string specifying the column name that contains the values to be normalized (e.g., "Abundance").

log2transform

Logical. If TRUE, log2 transformation is applied to the data before normalization. Default is TRUE.

normalize.bySubset

A character string specifying the column name to be used for subsetting the data before normalization. Default is NULL.

normalize.bySubset.matchTerm

A character string specifying the term to match within the subset column for normalization. Default is NULL.

normalize.bySubset.matchTerm.exactMatch

Logical. If TRUE, requires an exact match for the normalize.bySubset.matchTerm. Default is FALSE.

Value

A data.table containing the normalized data, with the same structure as the input dat.

Details

The median_normalisation function applies median normalization to the data, which adjusts the values in each sample such that the median value is consistent across samples. This is particularly useful in MS or other omics data where global shifts in signal intensities between samples need to be corrected.

The function allows for:

Log2 Transformation: Optional log2 transformation of the data prior to normalization.

Subset Normalization: Normalization based on a subset of the data, defined by the normalize.bySubset parameter.

If normalize.bySubset is provided, the function calculates the median based on the specified subset, otherwise, it normalizes the data using the entire dataset. The function also supports partial or exact matching for subsetting.

Examples

if (FALSE) { # \dontrun{
  data <- data.table(
    SampleID = c("S1", "S2", "S3", "S4"),
    Protein = c("P1", "P2", "P1", "P2"),
    Abundance = c(1.2, 2.3, 3.1, 4.2)
  )

  normalized_data <- median_normalisation(data, col.p = "SampleID", row.p = "Protein", val = "Abundance")
} # }