Merge Metadata with Mass Spectrometry Data

This function merges metadata with mass spectrometry data, ensuring consistency between the pools and channels in both datasets. It performs checks to validate the presence and consistency of Pool and Channel information between the datasets, and adjusts them as necessary to match.

Usage

merge_ms_metadata(msDT, metadata, extMixFrac, int.nm)

Arguments

msDT

A data.table containing the mass spectrometry data. This table should include columns such as Pool, Channel, and Protein.

metadata

A data.table containing the metadata to be merged with msDT. This table should include columns such as Pool and Channel.

extMixFrac

A data.table containing additional information for external mix fractions, used if Pool information is missing in the metadata.

int.nm

A character string specifying the internal name used for processing. (This parameter is referenced in the function but not used within the provided code.)

Value

A list containing two elements:

"msDT": The merged mass spectrometry data as a data.table.

"metadata": The merged metadata as a data.table.

Details

The merge_ms_metadata function ensures that the Pool and Channel columns in the mass spectrometry data (msDT) and the metadata are consistent. It checks for the presence of Pool information in both datasets, adjusts levels if necessary, and merges the metadata into the mass spectrometry data.

Specific steps include:

Verifying and correcting Pool levels between msDT and metadata.

Handling cases where Pool information is missing in either msDT or metadata.

Ensuring consistency of Channel information between the datasets.

Merging the metadata into msDT with consideration for potential inconsistencies.