Configure MSqb2 Analysis Parameters

This function sets up and configures parameters for an MSqb2 analysis. It allows users to specify various settings related to filtering, normalization, summarization, fraction combination, batch correction, and missing data imputation, among others.

Usage

msqb_config(
  build.para,
  config.file = NULL,
  interactive = TRUE,
  study.variable = "protein",
  filter.by.quaninfo = "auto",
  isolation.interference.cutoff = 75,
  collapse_psm_method = "ionscore",
  filter.singleshot.proteins = "ByPeptide",
  min.intensity = 0.01,
  normalization.method = "vsn",
  normalize.bySubset = NULL,
  summarization.method = "tmp",
  medianpolish.method = "column.eff",
  medianpolish.byPool = FALSE,
  fraction.collapse.method = "maxset",
  batch.corr.method = "none",
  na.imputation.method = "none",
  filter.metadata = NULL,
  feature.annotation.source = "psm.file",
  study.organism = NULL,
  pairwise.contrasts = TRUE,
  manual.contrasts = NULL,
  choose.contrasts = FALSE,
  pairwise.denominator = NULL,
  protein.column = "Master Protein Accessions",
  peptide.seq.column = "Sequence",
  peptide.annot.seq.column = "Annotated Sequence",
  charge.column = "Charge",
  FileID.column = "File ID",
  quaninfo.column = "Quan Info",
  isolation.interference.column = "Isolation Interference [%]",
  ion.score.column = "Ions Score",
  filename.column = "Spectrum File",
  abundance.columns = "Abundance",
  desc.column = "Master Protein Descriptions",
  contaminant.column = "Contaminant",
  marked.as.column = "Marked as",
  subsymDT = data.table(symout = c("\\~", "\\+", "\\:", "\\|", "\\-", "\\(",
    "\\)", "\\/", " "), symin = c("_", ".p", ".", ".", "_", ".", ".", ".", ".")),
  ...
)

Arguments

build.para

A list or object containing the build parameters for the MSqb2 analysis pipeline. This list is generated by the function MSqb_build().

config.file

A character string specifying the path to a configuration file to load. If NULL, no configuration file is loaded. Default is NULL.

interactive

Logical. If TRUE, the function may prompt the user for input during execution. Default is TRUE.

study.variable

A character string specifying the main study variable, either "protein" or "peptide". Default is "protein".

filter.by.quaninfo

A character string specifying how to filter data based on quantification information. Default is "auto".

isolation.interference.cutoff

Numeric. The cutoff value for isolation interference, used to filter data. Default is 75.

collapse_psm_method

A character string specifying the method for handling multiple features. Default is "ionscore".

filter.singleshot.proteins

A character string specifying how to filter single-shot proteins. Options are "ByPeptide" or "ByProtein". Default is "ByPeptide".

min.intensity

Numeric. The minimum intensity threshold for filtering data. Default is 0.01.

normalization.method

A character string specifying the normalization method. Options include "vsn", "quantile", "median", etc. Default is "vsn".

normalize.bySubset

A character string referring to a column name or NULL. If specified, normalization will be performed by the subset defined by the labels in the specified data column. Default is NULL.

summarization.method

A character string specifying the method for summarizing data. Options include "tmp", "median", "mean" and "sum" Default is "tmp".

medianpolish.method

A character string specifying the method for median polishing. Options include "column.eff" and "residual". Default is "column.eff".

medianpolish.byPool

Logical. If TRUE, median polishing is performed by pool (in TMT method). Default is FALSE.

fraction.collapse.method

A character string specifying the method for collapsing fractions. Options are "MAXset", "max", "median", "sum", "single", "multi". Default is "MAXset".

batch.corr.method

A character string specifying the method for batch correction. Options include "limma", "svn" and "none". Default is "none".

na.imputation.method

A character string specifying the method for imputing missing data. Options include all methods from packages msImpute and imputeQRILC. Default is "none".

filter.metadata

A character string or NULL. If specified, metadata will be filtered accordingly. For details see the help of filter_metadata() function. Default is NULL.

feature.annotation.source

A character string specifying the source of feature annotation. Options include "psm.file", "uniprot", "uniprot.file" and "none". Default is "psm.file".

study.organism

A character string specifying the organism studied. Only if feature.annotation.source is set to "uniprot". Default is NULL.

pairwise.contrasts

Logical. If TRUE, pairwise contrasts are automatically generated. Default is TRUE.

manual.contrasts

A list specifying manual contrasts to be applied. Default is NULL.

choose.contrasts

Logical. If TRUE, allows the user to choose contrasts interactively. Default is FALSE.

pairwise.denominator

A character string or NULL specifying the denominator for pairwise contrasts. Default is NULL.

protein.column

A character string specifying the column name for proteins in the input data. Default is "Master Protein Accessions".

peptide.seq.column

A character string specifying the column name for peptide sequences in the input data. Default is "Sequence".

peptide.annot.seq.column

A character string specifying the column name for annotated peptide sequences. Default is "Annotated Sequence".

charge.column

A character string specifying the column name for charge in the input data. Default is "Charge".

FileID.column

A character string specifying the column name for file IDs in the input data. Default is "File ID".

quaninfo.column

A character string specifying the column name for quantification information. Default is "Quan Info".

isolation.interference.column

A character string specifying the column name for isolation interference. Default is "Isolation Interference [%]".

ion.score.column

A character string specifying the column name for ion scores. Default is "Ions Score".

filename.column

A character string specifying the column name for filenames in the input data. Default is "Spectrum File".

abundance.columns

A character string specifying the column name for abundance values. Default is "Abundance".

desc.column

A character string specifying the column name for protein descriptions. Default is "Master Protein Descriptions".

contaminant.column

A character string specifying the column name for contaminants. Default is "Contaminant".

marked.as.column

A character string specifying the column name for marked entries in the input data. Default is "Marked as".

subsymDT

A data.table defining symbolic substitutions, with columns symout and symin for output and input symbols respectively.

...

Additional arguments passed to other methods or functions.

Value

A list or object containing the configured parameters for MSqb2 analysis.

Details

The msqb_config function provides a comprehensive interface for configuring an MSqb2 analysis. Users can specify various filtering criteria, normalization methods, summarization techniques, and other parameters necessary for processing proteomics data. The function is highly customizable and can be used interactively or with predefined configurations.