MSqb Workflow for Proteomics Data Analysis

This function orchestrates the entire workflow for MS proteomics data analysis, including data reformatting, normalization, batch effect removal, protein summarization, differential expression analysis, and more.

Usage

msqb_workflow(build.para, config.para = NULL, config.file = NULL, ...)

Arguments

build.para

A list of parameters required to build the workflow environment (see MSqb_build() function).

config.para

A list of configuration parameters for the workflow. If NULL, these will be read from the config.file (see MSqb_config() function).

config.file

The path to the configuration file. This file will be used if config.para is not provided.

...

Additional parameters passed to internal functions.

Value

An environment containing the results of the workflow, including processed data tables and statistical results.

Examples

if (FALSE) { # \dontrun{
result_env <- msqb_workflow(
  build.para = list(param1 = value1, param2 = value2),
  config.file = "config.R"
)
} # }