Normalize MS Data

This function normalizes mass spectrometry data using one of three available methods: Variance Stabilizing Normalization (VSN), median normalization, or quantile normalization. The function can also handle normalization by specific data subsets.

Usage

normalise_data(method = "VSN", path = Tables.path, ...)

Arguments

method

Character. The normalization method to use. Options are "vsn", "median", or "quantile". Default is "VSN".

path

Character. The path where the output Excel file with normalized and non-normalized data will be saved. Default is Tables.path.

...

Additional arguments passed to the normalization functions, including the column and row identifiers (col.p, row.p), the intensity column (val), and other specific parameters for each normalization method.

Value

A data.table containing the normalized data, with the intensity values replaced by the normalized abundance values.

Details

vsn: Variance Stabilizing Normalization is applied, with an option to normalize by a subset of the data.

median: Median normalization can be applied, with log2 transformation by default.

quantile: Quantile normalization is performed, transforming the intensity distributions of all samples to be identical.

The function also generates an Excel file containing both the normalized and non-normalized peptide-level data.

Examples

if (FALSE) { # \dontrun{
normalized_data <- normalise_data(
  method = "VSN",
  dat = my_data,
  val = "Intensity",
  path = "output_directory"
)
} # }