Read or Generate Configuration Parameters

Reads configuration parameters from a provided file or generates them based on build parameters. Handles cases where both configuration parameters and a configuration file are provided, prompting the user to choose which to use. The function can work for both workflow and visualization configuration types.

Usage

readConfigPara(build.para, config.para, config.file, config.type, ...)

Arguments

build.para

A list containing build parameters. This is generated by the msqb_build function.

config.para

A list containing pre-existing configuration parameters (optional).

config.file

A character string specifying the path to a configuration file (optional).

config.type

A character string indicating the type of configuration: either "wf" for workflow or "viz" for visualization.

...

Additional arguments passed to the configuration functions (msqb_config or msqb_config_viz).

Value

The function does not return a value directly. Instead, it modifies the environment by expanding and exporting the configuration parameters.

Details

This function is designed to handle the reading and generation of configuration parameters based on the build parameters and the specified configuration type. If both config.file and config.para are provided, the user is prompted to choose which to use. If neither is provided, the function attempts to generate the configuration parameters using the specified configuration function.

The function expands the configuration list and exports it to the top environment, making it available for further processing in the workflow.

Examples

if (FALSE) { # \dontrun{
# Assuming `build.para` is already defined
readConfigPara(build.para = build.para, config.file = "path/to/config.R", config.type = "wf")
} # }