Reformat PSM Data from Proteome Discoverer

This function reformats PSM-level data specifically generated from the Proteome Discoverer software, merging it with metadata and performing necessary clean-up steps for downstream analysis.

Usage

reformat_PD_data(
  feat.dt = feat.dt,
  metadata = metadata,
  protein.column = "Master Protein Accessions",
  peptide.seq.column = "Sequence",
  peptide.annot.seq.column = "Annotated Sequence",
  charge.column = "Charge",
  FileID.column = "File ID",
  ion.score.column = "Score",
  quaninfo.column = "Quan Info",
  isolation.interference.column = "Isolation Interference [%]",
  desc.column = "Master Protein Descriptions",
  contaminant.column = "Contaminant",
  marked.as.column = "Marked as",
  abundance.columns = "Abundance",
  ms.software = "PD",
  subsymDT = data.table(symout = c("\\~", "\\+", "\\:", "\\|", "\\-", "\\(",
    "\\)", "\\/", " "), symin = c("_", ".p", ".", ".", "_", ".", ".", ".", "."))
)

Arguments

feat.dt

A data.table containing PSM-level data from Proteome Discoverer.

metadata

A data.table containing metadata for the samples.

protein.column

Name of the column containing protein accession identifiers. Default is "Master Protein Accessions".

peptide.seq.column

Name of the column containing peptide sequences. Default is "Sequence".

peptide.annot.seq.column

Name of the column containing annotated peptide sequences. Default is "Annotated Sequence".

charge.column

Name of the column containing charge states. Default is "Charge".

FileID.column

Name of the column containing file IDs. Default is "File ID".

ion.score.column

Name of the column containing ion scores. Default is "Score".

quaninfo.column

Name of the column containing quantitation information. Default is "Quan Info".

isolation.interference.column

Name of the column containing isolation interference values. Default is "Isolation Interference %".

desc.column

Name of the column containing protein descriptions. Default is "Master Protein Descriptions".

contaminant.column

Name of the column indicating contaminants. Default is "Contaminant".

marked.as.column

Name of the column indicating marked entries. Default is "Marked as".

abundance.columns

Prefix for the columns containing abundance or intensity values. Default is "Abundance".

ms.software

Character string indicating the mass spectrometry software used. Default is "PD".

subsymDT

A data.table used to replace special characters in column names. Default is a predefined data.table.

Value

A list containing reformatted PSM data, metadata, and annotation data.

Examples

if (FALSE) { # \dontrun{
# Example usage of reformat_PD_data function for data from Proteome Discoverer
reformatted_data <- reformat_PD_data(feat.dt = my_data, metadata = my_metadata)
} # }