VSN Normalization

This function applies Variance Stabilizing Normalization (VSN) to mass spectrometry data, with options to normalize based on subsets of the data or the entire dataset.

Usage

vsn_normalisation(
  dat,
  col.p,
  row.p,
  val,
  calib = "affine",
  normalize.bySubset = NULL,
  normalize.bySubset.matchTerm = NULL,
  normalize.bySubset.matchTerm.exactMatch = FALSE,
  sp = "_.._"
)

Arguments

dat

A data.table containing the data to be normalized.

col.p

Character vector. Column names used for grouping the data (e.g., sample or condition).

row.p

Character vector. Row identifiers, typically the feature or protein IDs.

val

Character. The column name representing the intensity or abundance values to be normalized.

calib

Character. Calibration method to use, either "affine" (default) or "none". If "none", it assumes the data is already normalized.

normalize.bySubset

Character or NULL. Column name used to define a subset for normalization. If NULL, the entire dataset is normalized.

normalize.bySubset.matchTerm

Character or NULL. The term used to match entries within the subset defined by normalize.bySubset. If NULL, the whole dataset is used.

normalize.bySubset.matchTerm.exactMatch

Logical. If TRUE, only exact matches of normalize.bySubset.matchTerm are used for subsetting. Default is FALSE.

sp

Character. Separator used for pasting together row identifiers when converting data between wide and long formats. Default is "_.._".

Value

A data.table with normalized intensity values.

Details

This function uses the vsn2 function from the vsn package to perform normalization. The normalization can be applied to the entire dataset or to a specific subset, which can be useful if different subsets require separate normalization.

Subset Normalization: If normalize.bySubset is specified, only the subset matching normalize.bySubset.matchTerm is used for normalization.

Calibration: The calib parameter allows for affine calibration or no calibration ("none").

Examples

if (FALSE) { # \dontrun{
  normalized_data <- vsn_normalisation(dat = my_data, col.p = "SampleID", row.p = "ProteinID", val = "Intensity")
} # }